# contracts/openapi.json
{
  "components": {
    "schemas": {
      "BrowseFoldersIn": {
        "properties": {
          "auth_mode": {
            "enum": [
              "service_account",
              "oauth"
            ],
            "title": "Auth Mode",
            "type": "string"
          },
          "credential_ref": {
            "minLength": 1,
            "title": "Credential Ref",
            "type": "string"
          },
          "parent_id": {
            "default": "root",
            "title": "Parent Id",
            "type": "string"
          },
          "principal": {
            "minLength": 1,
            "title": "Principal",
            "type": "string"
          }
        },
        "required": [
          "auth_mode",
          "principal",
          "credential_ref"
        ],
        "title": "BrowseFoldersIn",
        "type": "object"
      },
      "BrowseFoldersOut": {
        "properties": {
          "folders": {
            "items": {
              "$ref": "#/components/schemas/DriveFolderOut"
            },
            "title": "Folders",
            "type": "array"
          },
          "parent_id": {
            "title": "Parent Id",
            "type": "string"
          }
        },
        "required": [
          "parent_id",
          "folders"
        ],
        "title": "BrowseFoldersOut",
        "type": "object"
      },
...
      "CredentialIn": {
        "description": "A source's credential. Write-only -- there is no read endpoint.",
        "properties": {
          "secret": {
            "description": "Service-account JSON key, or a JSON object with refresh_token, client_id and client_secret for OAuth.",
            "minLength": 1,
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret"
        ],
        "title": "CredentialIn",
        "type": "object"
      },
...
      "DriveFolderOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "DriveFolderOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoredCredentialOut": {
        "description": "The result of storing a credential outside any source (ADR-0016).\n\nDecoupled from a source id on purpose: a credential can now be provisioned\n-- by pasting a key or by completing OAuth -- *before* the source it will\nbelong to has been created or saved, which is what makes browsing folders\nfrom the create dialog possible instead of requiring save-then-edit.",
        "properties": {
          "credential_ref": {
            "title": "Credential Ref",
            "type": "string"
          }
        },
        "required": [
          "credential_ref"
        ],
        "title": "StoredCredentialOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nRAG chat over Google Drive and other document sources.\n\n**Retrieval** is hybrid: dense pgvector search and Postgres full-text search are\nfused with Reciprocal Rank Fusion, then reranked (ADR-0004). Every turn streams\nits retrieval trace so a wrong answer can be diagnosed rather than merely\nnoticed (ADR-0009).\n\n**Permissions** are respected during ingestion. A folder that cannot be read is\nskipped *with an audit record* and never fails the run, because a silent skip is\nindistinguishable from an empty folder (ADR-0003).\n",
    "title": "RAGDrive API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/credentials": {
      "post": {
        "description": "Store a secret, encrypted at rest, under a freshly generated reference.\n\nThe reference is always server-generated, never typed by a user -- a\nhand-typed `credential_ref` is one typo from silently pointing a source at\nnothing, or at another source's key.",
        "operationId": "storeCredential",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredCredentialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Store Credential",
        "tags": [
          "credentials"
        ]
      }
    },
...
    "/oauth/google/callback": {
      "get": {
        "description": "Exchange the code, store the credential, send the browser home.\n\nEvery failure path bounces back into the frontend with an error message in\nthe query string rather than rendering a raw API error page -- the browser\narrives here mid-navigation from Google, not from a fetch call the frontend\ncan catch, so an API-shaped JSON error is not something a user would ever\nusefully see.",
        "operationId": "googleOAuthCallback",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          },
          {
            "in": "cookie",
            "name": "ragdrive_oauth_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ragdrive Oauth State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback",
        "tags": [
          "oauth"
        ]
      }
    },
    "/oauth/google/start": {
      "get": {
        "description": "Redirect the browser to Google's consent screen.\n\n`return_path` and `editing_source_id` exist so the *same* button and flow\nwork from both the create dialog (no source yet) and an existing source's\nedit dialog (reconnecting or replacing its credential): both round-trip\nthrough `state` and come back on the callback redirect.",
        "operationId": "startGoogleOAuth",
        "parameters": [
          {
            "in": "query",
            "name": "return_path",
            "required": false,
            "schema": {
              "default": "/configuration",
              "title": "Return Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "editing_source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Editing Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start",
        "tags": [
          "oauth"
        ]
      }
    },
...
    "/sources/browse-folders": {
      "post": {
        "description": "List the subfolders of a Drive folder, for the config UI's folder picker.\n\nTakes a `credential_ref` directly rather than a source id, so this works\nfrom the create dialog too -- before any source row exists, right after\nOAuth connects or a service-account key is stored via `POST /credentials`.",
        "operationId": "browseFolders",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrowseFoldersIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseFoldersOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Browse Folders",
        "tags": [
          "sources"
        ]
      }
    },
...
    "/sources/{source_id}/credential": {
      "put": {
        "description": "Store a source's credential, encrypted at rest.\n\nWrite-only by design. There is no GET: a credential that can be read back\nover HTTP is a credential one misconfigured permission away from disclosure,\nand nothing in the product needs to display it.",
        "operationId": "setSourceCredential",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Credential",
        "tags": [
          "sources"
        ]
      }
    },
...
  }
}